dox and giving root account access to everything by default
And some cleaning up. Also added override of [p5.auth.users.edit] that
allows for keeping old settings.

+++ plugins/extras/p5.auth/Misc.cs
@@ -21,8 +21,10 @@
  * out our website at http://gaiasoul.com for more details.
  */
 
+using p5.exp;
 using p5.core;
 using p5.auth.helpers;
+using p5.exp.exceptions;
 
 namespace p5.auth
 {
@@ -68,6 +70,8 @@ namespace p5.auth
         [ActiveEvent (Name = "p5.auth.misc.delete-my-user")]
         public static void p5_auth_misc_delete_my_user (ApplicationContext context, ActiveEventArgs e)
         {
+            if (context.Ticket.Role == "root")
+                throw new LambdaException ("You can't delete a root account, without making it a non-root account first", e.Args, context);
             using (new ArgsRemover (e.Args, true)) {
                 AuthenticationHelper.DeleteMyUser (context, e.Args);
             }
